Capture the pointer
This fixes a lot of edge case behavior. For example, minimize the browser, start dragging the bar, and then move the cursor outside of the browser. The bar will stop dragging (bug!). Then release the mouse outside of the window and move the cursor back into the browser. The bar will now start following the cursor again, even though it is not dragging (bug!).

Capturing the pointer solves both of these issues.