Main application structure (#21)
* Implemented the sidebars and made them resizable

Co-Authored-By: Nyby <45389841+NybyDK@users.noreply.github.com>

* Implemented navigation bars

Created the skeleton of the navigation bars and made the sidebars and canvas resizable.

Big thanks to @NybyDK! Without their knowledge this would not be possible!

Co-Authored-By: Nyby <45389841+NybyDK@users.noreply.github.com>

* Optimized the adding and removing of eventlisteners

* Uncommented the outdated tests and cleaned up

* Fixed the GUI test

* Simplify resize event

It only runs when `currentSide` has a value, otherwise it is unregistered.

* Prevent default on click

This ensures that text is not selected when moving the bar around

* Capture the pointer

This fixes a lot of edge case behavior. For example, minimize the browser, start dragging the bar, and then move the cursor outside of the browser. The bar will stop dragging (bug!). Then release the mouse outside of the window and move the cursor back into the browser. The bar will now start following the cursor again, even though it is not dragging (bug!).

Capturing the pointer solves both of these issues.

* Listen for pointer cancel

If we don't listen for this, there can be edge cases where we never stop listening for a cursor that no longer exists. For example if the mouse loses connection while dragging.

* Clarified some function and variable names and wrote some comments

Co-Authored-By: Nyby <45389841+NybyDK@users.noreply.github.com>

* Attempted to streamline the variable, function and Enum name

* Implemented a more scalable solution

Implemented a more scalable solution to the sidepanels, top- and bottom navbars.

* Discarded the percentage solution

* Removed unused import

---------

Co-authored-by: Nyby <45389841+NybyDK@users.noreply.github.com>
Co-authored-by: Anton <dev@atjn.dk>